Use QImageReader instead of QPixmap constructor to load images.
This is much more efficient, as it doesn't insert items into QPixmapCache.

This would normally perhaps not be such a bad thing, but we expect QML to be
'smart' enough to do the caching for us, so this just creates unnecessary cache
pressure and wastes memory.

This also avoids an unnecessary check on the file mtime due to not using the
cache, meaning less unnecessary disk access.

This results in ~1mb less VSZ and ~1mb less RES memory usage on a Lenovo S10-3t,
and apparently faster startup.

Signed-off-by: Robin Burchell <viroteck@viroteck.net>
Signed-off-by: John Brooks <john.brooks@dereferenced.net>