Fixes bug 20077
The Problem was that the focus was relayed to the BorderImage (root item
of the TextEntry) which reacted to focus changed by relaying the focus
to teh TextInput. Unfortunately by relaying it it also relayed losing
the focus.

A tap on the GestureArea (which fills the whole TextEntry) now relays
the focus directly to the TextInput instead of the surrounding
BorderImage, so the behavior is fixed for tapping.