Fixes bug 20119
The TextEntry is now enclosed in a FocusScope which works as a focus proxy and relays the focus changes to the child element that has the property "focus" set to true (in the TextEntry that is the TextInput element), thus giving it activeFocus.
onFocusChanged has been removed as it triggered the removal of the focus
from the input as soon as it had relayed the focus to the input because
losing the focus also triggered onFocusChanged.
The solution replaces the fix for bugs 20077 and 20022.